+++ fuzz/fuzz_targets/addrv2_wire.rs
@@ -1,0 +1,7 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = rbitcoin_net::parse_v2_regtest_named("addrv2", data);
+});